Feat : LuckyGame 클래스 만들기
